todo-app/requirements.txt
djangorestframework
psycopg2-binary
python-decouple # для скрытия данных в .env
